pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["config*", "schemas*", "locales*", "utils*", "tools*", "workflows*", "api*", "frontend*"]

[project]
name = "magna-ai-lc-platform"
version = "2.0.0"
description = "L/C Document Processing — FastMCP + LangGraph + FastAPI"
requires-python = ">=3.11"

dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",

    # MCP (FastMCP)
    "fastmcp>=2.14",

    # Orchestration (LangGraph)
    "langgraph>=1.0",

    # API Server (FastAPI)
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",

    # LLM Providers
    "google-genai>=1.0",
    "openai>=1.30",

    # PDF
    "PyPDF2>=3.0",
    "Pillow>=10.0",

    # External APIs
    "httpx>=0.27",

    # Database
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
]

[project.optional-dependencies]
ocr = ["pytesseract", "pdf2image"]
dev = ["pytest", "pytest-asyncio", "ruff", "httpx"]
streamlit = ["streamlit>=1.35"]
